MyRESTaurantAPI-python/Menu/GetEmotions/requirements.txt
Werkzeug==2.0.3
functions_framework==3.3.0
google-cloud-language==2.9.0